Increased the size of handlable problem data from 255 to 4294967295.

+++ rawFunction.m
@@ -5,17 +5,17 @@ THEFOLDER = 'REPLACE_PATH';
 TEMP = [bin,-Ain];
 
 fid = fopen('/Volumes/MyDrive/MATLABtoLOWLEVEL','w+');
-fwrite(fid, uint8(size(TEMP,1)), 'uint8');
-fwrite(fid, uint8(size(TEMP,2)), 'uint8');
+fwrite(fid, uint32(size(TEMP,1)), 'uint32');
+fwrite(fid, uint32(size(TEMP,2)), 'uint32');
 fwrite(fid, TEMP, 'double');
-fwrite(fid, uint8(LAST_ARGUMENT), 'uint8');
+fwrite(fid, uint32(LAST_ARGUMENT), 'uint32');
 fclose(fid);
 
 [~,~] = system([THEFOLDER,'/FUNCTION_CALL'],'-echo');
 
 fid = fopen('/Volumes/MyDrive/LOWLEVELtoMATLAB','r');
-m = fread(fid,1,'uint8');
-n = fread(fid,1,'uint8');
+m = fread(fid,1,'uint32');
+n = fread(fid,1,'uint32');
 out = fread(fid, [m, n], 'double');
 fclose(fid);
 [~,~] = system('rm /Volumes/MyDrive/MATLABtoLOWLEVEL /Volumes/MyDrive/LOWLEVELtoMATLAB');
